+++ src/monoazul/duckdbffi/driver/Result.java
@@ -1,6 +1,5 @@
 package monoazul.duckdbffi.driver;
 
-import monoazul.duckdbffi.jextractffi.duckdb_h;
 import monoazul.duckdbffi.jextractffi.duckdb_result;
 
 import java.lang.foreign.Arena;
@@ -9,7 +8,6 @@ import java.lang.foreign.ValueLayout;
 import java.util.ArrayList;
 import java.util.List;
 
-// FFI Imports
 import static monoazul.duckdbffi.jextractffi.duckdb_h.*;
 
 public class Result
@@ -42,20 +40,22 @@ public class Result
             MemorySegment DuckDbResult = duckdb_result.reinterpret(DuckDbResultPtr, ResulArena, null);
             MemorySegment DbChunk = duckdb_fetch_chunk(DuckDbResult);
 
-            if (DbChunk.address() == 0) {
+            if (DbChunk.address() == 0)
+            {
                 // Empty result
                 ResMetaData = new ResultMetaData(0, columnsCount, 0, maxVectorSize);
                 return;
             }
             int chunkCount = 1;
-            int dbChunkSize = (int) duckdb_data_chunk_get_size(DbChunk);
+            int dbChunkSize = (int)duckdb_data_chunk_get_size(DbChunk);
             int rowCount = dbChunkSize;
 
             // Create subclasses of Column for all columns
-            for (int col = 0; col < columnsCount; col++) {
+            for (int col = 0; col < columnsCount; col++)
+            {
                 MemorySegment ResultVector = duckdb_data_chunk_get_vector(DbChunk, col);
                 MemorySegment ResultVectorType = duckdb_vector_get_column_type(ResultVector);
-                DuckDbDatatype DbDatatype = new DuckDbDatatype((short) duckdb_get_type_id(ResultVectorType));
+                DuckDbDatatype DbDatatype = new DuckDbDatatype((short)duckdb_get_type_id(ResultVectorType));
                 String ColumnName = duckdb_column_name(DuckDbResultPtr, col).reinterpret(Integer.MAX_VALUE).getString(0);
 
                 // Destroy column_type, but we need a pointer first
@@ -78,12 +78,14 @@ public class Result
             // Fill Columns chunk-wise
             DbChunk = duckdb_fetch_chunk(DuckDbResult);
 
-            while (DbChunk.address() != 0) {
-                dbChunkSize = (int) duckdb_data_chunk_get_size(DbChunk);
+            while (DbChunk.address() != 0)
+            {
+                dbChunkSize = (int)duckdb_data_chunk_get_size(DbChunk);
                 rowCount = +dbChunkSize;
                 chunkCount++;
 
-                for (int col = 0; col < columnsCount; col++) {
+                for (int col = 0; col < columnsCount; col++)
+                {
                     MemorySegment ResultVector = duckdb_data_chunk_get_vector(DbChunk, col);
                     Columns.get(col).addVectorChunk(ResultVector, dbChunkSize);
                 }
@@ -95,7 +97,8 @@ public class Result
 
             ResMetaData = new ResultMetaData(rowCount, columnsCount, chunkCount, maxVectorSize);
             // Add Result Metadata to Columns
-            for (int col = 0; col < columnsCount; col++) {
+            for (int col = 0; col < columnsCount; col++)
+            {
                 Columns.get(col).addResultMetaData(ResMetaData);
             }
         }
@@ -104,15 +107,14 @@ public class Result
     private static Column createColumnByDatatype(String ColumnName, DuckDbDatatype DbDatatype, boolean primitivesAsObject)
     {
         System.out.println(DbDatatype.type);
-        return switch (DbDatatype.type) {
+        return switch (DbDatatype.type)
+        {
             case DuckDbDatatype.DUCKDB_TYPE_INTEGER ->
                     ColumnFactory.createIntColumn(ColumnName, DbDatatype, primitivesAsObject);
             case DuckDbDatatype.DUCKDB_TYPE_BIGINT ->
                     ColumnFactory.createLongColumn(ColumnName, DbDatatype, primitivesAsObject);
-            case DuckDbDatatype.DUCKDB_TYPE_VARCHAR ->
-                    new StringColumn(ColumnName, DbDatatype);
-            case DuckDbDatatype.DUCKDB_TYPE_DECIMAL ->
-                    new DecimalColumn(ColumnName, DbDatatype);
+            case DuckDbDatatype.DUCKDB_TYPE_VARCHAR -> new StringColumn(ColumnName, DbDatatype);
+            case DuckDbDatatype.DUCKDB_TYPE_DECIMAL -> new DecimalColumn(ColumnName, DbDatatype);
             case DuckDbDatatype.DUCKDB_TYPE_FLOAT ->
                     ColumnFactory.createFloatColumn(ColumnName, DbDatatype, primitivesAsObject);
             case DuckDbDatatype.DUCKDB_TYPE_DOUBLE ->
@@ -124,39 +126,30 @@ public class Result
             case DuckDbDatatype.DUCKDB_TYPE_TIMESTAMP,
                  DuckDbDatatype.DUCKDB_TYPE_TIMESTAMP_S,
                  DuckDbDatatype.DUCKDB_TYPE_TIMESTAMP_MS,
-                 DuckDbDatatype.DUCKDB_TYPE_TIMESTAMP_NS ->
-                    new LocalDateTimeColumn(ColumnName, DbDatatype);
+                 DuckDbDatatype.DUCKDB_TYPE_TIMESTAMP_NS -> new LocalDateTimeColumn(ColumnName, DbDatatype);
             case DuckDbDatatype.DUCKDB_TYPE_TIMESTAMP_TZ ->
-                    // It is possible to use a factory in the future to return e.g. ZonedDateTime
+                // It is possible to use a factory in the future to return e.g. ZonedDateTime
                     new InstantColumn(ColumnName, DbDatatype);
-            case DuckDbDatatype.DUCKDB_TYPE_USMALLINT ->
-                    new IntObjectColumn(ColumnName, DbDatatype);
-            case DuckDbDatatype.DUCKDB_TYPE_UINTEGER->
-                    new LongObjectColumn(ColumnName, DbDatatype);
-            case DuckDbDatatype.DUCKDB_TYPE_UTINYINT ->
-                    new ShortObjectColumn(ColumnName, DbDatatype);
+            case DuckDbDatatype.DUCKDB_TYPE_USMALLINT -> new IntObjectColumn(ColumnName, DbDatatype);
+            case DuckDbDatatype.DUCKDB_TYPE_UINTEGER -> new LongObjectColumn(ColumnName, DbDatatype);
+            case DuckDbDatatype.DUCKDB_TYPE_UTINYINT -> new ShortObjectColumn(ColumnName, DbDatatype);
             case DuckDbDatatype.DUCKDB_TYPE_HUGEINT,
                  DuckDbDatatype.DUCKDB_TYPE_UHUGEINT,
-                 DuckDbDatatype.DUCKDB_TYPE_UBIGINT->
-                    new BigIntegerColumn(ColumnName, DbDatatype);
+                 DuckDbDatatype.DUCKDB_TYPE_UBIGINT -> new BigIntegerColumn(ColumnName, DbDatatype);
             case DuckDbDatatype.DUCKDB_TYPE_BOOLEAN ->
                     ColumnFactory.createBooleanColumn(ColumnName, DbDatatype, primitivesAsObject);
-            case DuckDbDatatype.DUCKDB_TYPE_DATE ->
-                    new LocalDateColumn(ColumnName, DbDatatype);
+            case DuckDbDatatype.DUCKDB_TYPE_DATE -> new LocalDateColumn(ColumnName, DbDatatype);
             case DuckDbDatatype.DUCKDB_TYPE_TIME,
-                 DuckDbDatatype.DUCKDB_TYPE_TIME_NS ->
-                    new LocalTimeColumn(ColumnName, DbDatatype);
-            case DuckDbDatatype.DUCKDB_TYPE_UUID ->
-                    new UuidColumn(ColumnName, DbDatatype);
-            case DuckDbDatatype.DUCKDB_TYPE_INTERVAL ->
-                    new IntervalColumn(ColumnName, DbDatatype);
+                 DuckDbDatatype.DUCKDB_TYPE_TIME_NS -> new LocalTimeColumn(ColumnName, DbDatatype);
+            case DuckDbDatatype.DUCKDB_TYPE_UUID -> new UuidColumn(ColumnName, DbDatatype);
+            case DuckDbDatatype.DUCKDB_TYPE_INTERVAL -> new IntervalColumn(ColumnName, DbDatatype);
             default -> new UnknownColumn(ColumnName, DbDatatype);
         };
     }
 
     public List<Object> getRow(int row)
     {
-        List<Object> Row = new ArrayList<Object>((int)ResMetaData.columnsCount());
+        List<Object> Row = new ArrayList<Object>(ResMetaData.columnsCount());
         for (int col = 0; col < ResMetaData.columnsCount(); col++)
         {
             Row.add(Columns.get(col).getValue(row));
